backend/requirements.txt
# backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
django-storages==1.14.2
boto3==1.34.34
bleach==6.1.0